requirements.txt
numpy>=1.23
opencv-python>=4.7
torch>=2.0
torchvision>=0.15
facenet-pytorch>=2.5.2
ultralytics>=8.0.20
Pillow>=10.0
PyYAML>=6.0
scikit-learn>=1.3
mysql-connector-python>=8.0
torchvision
tensorflow==1.15  # for TF1-style code and BeautyGAN/DiscoFaceGAN
scipy
